Update spatiotemporal diffusion modules
Move multichannel advection-diffusion and the config-backed
reaction-diffusion implementation into the canonical spatiotemporal
namespace. Keep legacy module paths as wrappers and add a numerical
comparison test against the previous stable reaction implementation.
