+++ docs/_toc.yml
@@ -22,21 +22,20 @@ chapters:
           - file: reference/simulations/base
           - file: reference/simulations/spatiotemporal/index
             sections:
+              - file: reference/simulations/spatiotemporal/advection_diffusion_multichannel
               - file: reference/simulations/spatiotemporal/gray_scott
               - file: reference/simulations/spatiotemporal/gross_pitaevskii
               - file: reference/simulations/spatiotemporal/navier_stokes_conditioned
+              - file: reference/simulations/spatiotemporal/reaction_diffusion
           - file: reference/simulations/advection_diffusion
-          - file: reference/simulations/advection_diffusion_multichannel
           - file: reference/simulations/double_pendulum
           - file: reference/simulations/epidemic
           - file: reference/simulations/flow_problem
           - file: reference/simulations/projectile
-          - file: reference/simulations/reaction_diffusion
           - file: reference/simulations/seir
       - file: reference/experimental/index
         sections:
           - file: reference/experimental/compressible_fluid
           - file: reference/experimental/hydrodynamics_2d
           - file: reference/experimental/lattice_boltzmann
-          - file: reference/experimental/reaction_diffusion
           - file: reference/experimental/shallow_water
