--- docs/reference/experimental/reaction_diffusion.rst
@@ -1,7 +1,0 @@
-autosim.experimental.simulations.reaction_diffusion
-===================================================
-
-.. automodule:: autosim.experimental.simulations.reaction_diffusion
-   :members:
-   :undoc-members:
-   :show-inheritance:
