--- docs/reference/simulations/advection_diffusion_multichannel.rst
@@ -1,7 +1,0 @@
-autosim.simulations.advection_diffusion_multichannel
-====================================================
-
-.. automodule:: autosim.simulations.advection_diffusion_multichannel
-   :members:
-   :undoc-members:
-   :show-inheritance:
