+++ docs/reference/simulations/spatiotemporal/advection_diffusion_multichannel.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.advection_diffusion_multichannel
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.advection_diffusion_multichannel
+   :members:
+   :undoc-members:
+   :show-inheritance:
