+++ docs/reference/simulations/spatiotemporal/reaction_diffusion.rst
@@ -1,0 +1,7 @@
+autosim.simulations.spatiotemporal.reaction_diffusion
+======================================================================
+
+.. automodule:: autosim.simulations.spatiotemporal.reaction_diffusion
+   :members:
+   :undoc-members:
+   :show-inheritance:
