+++ src/autosim/configs/simulator/advection_diffusion_multichannel.yaml
@@ -1,4 +1,4 @@
-_target_: autosim.simulations.AdvectionDiffusionMultichannel
+_target_: autosim.simulations.spatiotemporal.AdvectionDiffusionMultichannel
 output_indices: [0, 1, 2, 3]
 return_timeseries: true
 log_level: warning
