+++ src/autosim/configs/simulator/reaction_diffusion.yaml
@@ -1,4 +1,4 @@
-_target_: autosim.experimental.simulations.ReactionDiffusion
+_target_: autosim.simulations.spatiotemporal.ReactionDiffusion
 return_timeseries: true
 log_level: warning
 n: 64
