+++ src/autosim/simulations/advection_diffusion_multichannel.py
@@ -1,381 +1,3 @@
-"""Advection-diffusion simulator with multi-channel outputs.
+"""Compatibility wrapper for the spatiotemporal advection-diffusion simulator."""
 
-Returns channels: [vorticity, u, v, streamfunction].
-
-Implementation uses spectral Poisson solve (FFT) for streamfunction and central finite
-differences (via numpy.roll) for spatial derivatives.
-
-"""
-
-from __future__ import annotations
-
-import numpy as np
-import torch
-from scipy.integrate import solve_ivp
-
-from autosim.simulations.base import SpatioTemporalSimulator
-from autosim.types import NumpyLike, TensorLike
-
-# Integrator settings
-integrator_keywords = {"rtol": 1e-6, "atol": 1e-8, "method": "RK45"}
-
-
-class AdvectionDiffusionMultichannel(SpatioTemporalSimulator):
-    r"""Differentiable advection-diffusion simulator exposing multi-channel outputs.
-
-    The simulator evolves a vorticity field according to:
-
-    .. math::
-
-        \begin{aligned}
-        \partial_t \omega
-            &= \nu \nabla^2 \omega
-            - \mu (u \partial_x \omega + v \partial_y \omega)
-        \end{aligned}
-
-    It returns selected channels from
-    :math:`[\omega, u, v, \psi]`, where :math:`\psi` is the streamfunction.
-
-    Args:
-        parameters_range: Bounds on the sampled viscosity (`nu`) and advection strength
-            (`mu`).
-        output_names: Human-readable names for the returned channels.
-        output_indices: Channel indices to keep from ``[vorticity, u, v,
-            streamfunction]``. Defaults to all channels in canonical order.
-        return_timeseries: Whether `forward` returns the entire trajectory instead of a
-            single snapshot.
-        log_level: Logging verbosity passed to the base `Simulator`.
-        n: Number of spatial points per dimension.
-        L: Domain length in each spatial direction.
-        T: Total integration time.
-        dt: Temporal resolution used for the ODE solver outputs.
-        integrator_kwargs: Extra keyword arguments forwarded to
-            `scipy.integrate.solve_ivp`.
-
-    Notes:
-        Each grid point emits four channels `[vorticity, u, v, streamfunction]`.
-    """
-
-    _ALL_CHANNEL_NAMES = ("vorticity", "u", "v", "streamfunction")
-
-    def __init__(
-        self,
-        parameters_range: dict[str, tuple[float, float]] | None = None,
-        output_names: list[str] | None = None,
-        output_indices: list[int] | None = None,
-        return_timeseries: bool = False,
-        log_level: str = "progress_bar",
-        n: int = 32,
-        L: float = 10.0,
-        T: float = 80.0,
-        dt: float = 0.25,
-        integrator_kwargs: dict | None = None,
-    ) -> None:
-        """Initialize the multi-channel advection-diffusion simulator."""
-        if parameters_range is None:
-            parameters_range = {"nu": (0.0001, 0.01), "mu": (0.5, 2.0)}
-
-        if output_indices is None:
-            output_indices = [0, 1, 2, 3]
-        if len(output_indices) == 0:
-            msg = "output_indices must contain at least one channel index."
-            raise ValueError(msg)
-        if len(set(output_indices)) != len(output_indices):
-            msg = "output_indices must not contain duplicate channel indices."
-            raise ValueError(msg)
-
-        invalid_indices = [idx for idx in output_indices if idx < 0 or idx > 3]
-        if invalid_indices:
-            msg = (
-                "output_indices entries must be in the range [0, 3] for channels "
-                "[vorticity, u, v, streamfunction]. "
-                f"Received invalid indices: {invalid_indices}."
-            )
-            raise ValueError(msg)
-
-        selected_output_names = [self._ALL_CHANNEL_NAMES[idx] for idx in output_indices]
-        if output_names is None:
-            output_names = selected_output_names
-        elif len(output_names) != len(output_indices):
-            msg = (
-                "output_names length must match selected output_indices length. "
-                f"Received {len(output_names)} names for {len(output_indices)} indices."
-            )
-            raise ValueError(msg)
-
-        super().__init__(parameters_range, output_names, log_level)
-
-        self.output_indices = output_indices
-        self.return_timeseries = return_timeseries
-        self.n = n
-        self.L = L
-        self.T = T
-        self.dt = dt
-        self.integrator_kwargs = {**integrator_keywords, **(integrator_kwargs or {})}
-
-    def _forward(self, x: TensorLike) -> TensorLike:
-        # Expect single input sample in batch
-        assert x.shape[0] == 1, (
-            f"Simulator._forward expects a single input, got {x.shape[0]}"
-        )
-
-        # x contains the physical parameters [nu, mu]
-        sample = x.cpu().numpy()[0]
-
-        sol = simulate_advection_diffusion(
-            sample,
-            self.return_timeseries,
-            self.n,
-            self.L,
-            self.T,
-            self.dt,
-            self.integrator_kwargs,
-        )
-
-        # sol shape: (nt, n, n, channels) if return_timeseries else (n, n, channels)
-        arr = np.asarray(sol, dtype=np.float32)
-
-        # Flatten to (1, -1) for compatibility with the superclass
-        return torch.from_numpy(arr.ravel()).reshape(1, -1)
-
-    def forward_samples_spatiotemporal(
-        self,
-        n: int,
-        random_seed: int | None = None,
-        ensure_exact_n: bool = False,
-    ) -> dict:
-        """Produce simulator rollouts along with the sampled parameters.
-
-        Args:
-            n: Number of trajectories to sample.
-            random_seed: Seed for reproducible parameter draws.
-            ensure_exact_n: Whether to resample failed trajectories until exactly ``n``
-                succeed.
-
-        Returns:
-            Spatiotemporal payload with ``data``, ``constant_scalars``, and
-            ``constant_fields`` keys. Channels follow the configured
-            ``output_indices`` order.
-        """
-        y, x = self._forward_batch_with_optional_retries(
-            n=n,
-            random_seed=random_seed,
-            ensure_exact_n=ensure_exact_n,
-        )
-
-        channels = 4
-        features_per_step = self.n * self.n * channels
-        if self.return_timeseries:
-            total_features = y.shape[1]
-            if total_features % features_per_step != 0:
-                raise RuntimeError(
-                    "Returned tensor does not align with n*n*channels; "
-                    f"received {total_features} features, expected multiples of "
-                    f"{features_per_step}."
-                )
-            n_time = total_features // features_per_step
-            y_reshaped = y.reshape(y.shape[0], n_time, self.n, self.n, channels)
-        else:
-            if y.shape[1] != features_per_step:
-                raise RuntimeError(
-                    "Unexpected flattened size for single snapshot; "
-                    f"received {y.shape[1]}, expected {features_per_step}."
-                )
-            y_reshaped = y.reshape(y.shape[0], 1, self.n, self.n, channels)
-
-        # Select configured subset of channels in user-specified order.
-        y_reshaped = y_reshaped[..., self.output_indices]
-
-        return {
-            "data": y_reshaped,
-            "constant_scalars": x,
-            "constant_fields": None,
-        }
-
-
-def _spectral_poisson_solver(w2d: np.ndarray, K3: np.ndarray) -> np.ndarray:
-    r"""Solve ``laplacian(psi) = -omega`` in Fourier space.
-
-    Args:
-        w2d: Two-dimensional vorticity field with shape ``(n, n)``.
-        K3: Pre-computed spectral multiplier ``1 / (k_x^2 + k_y^2)`` with the zero mode
-            handled to keep the mean streamfunction at zero.
-
-    Returns:
-        Real-valued streamfunction with shape ``(n, n)``.
-    """
-    psi_hat = np.fft.fft2(w2d) * K3
-    psi = np.real(np.fft.ifft2(psi_hat))
-    return psi
-
-
-def _laplacian_periodic(w2d: np.ndarray, dx: float) -> np.ndarray:
-    """Apply the periodic Laplacian using second-order central differences.
-
-    Args:
-        w2d: Input field of shape ``(n, n)``.
-        dx: Grid spacing.
-
-    Returns:
-        Field after applying the Laplacian, matching the input shape.
-    """
-    return (np.roll(w2d, -1, axis=0) - 2 * w2d + np.roll(w2d, 1, axis=0)) / dx**2 + (
-        np.roll(w2d, -1, axis=1) - 2 * w2d + np.roll(w2d, 1, axis=1)
-    ) / dx**2
-
-
-def _gradient_periodic(f2d: np.ndarray, dx: float) -> tuple[np.ndarray, np.ndarray]:
-    """Compute periodic central-difference gradients.
-
-    Args:
-        f2d: Scalar field of shape ``(n, n)``.
-        dx: Grid spacing.
-
-    Returns:
-        ``(df/dx, df/dy)`` each with shape ``(n, n)``.
-    """
-    dfdx = (np.roll(f2d, -1, axis=1) - np.roll(f2d, 1, axis=1)) / (2 * dx)
-    dfdy = (np.roll(f2d, -1, axis=0) - np.roll(f2d, 1, axis=0)) / (2 * dx)
-    return dfdx, dfdy
-
-
-def advection_diffusion_rhs(
-    _t: float,
-    w_flat: np.ndarray,
-    n: int,
-    dx: float,
-    nu: float,
-    mu: float,
-    K3: np.ndarray,
-) -> np.ndarray:
-    r"""Evaluate the vorticity time derivative.
-
-    Args:
-        _t: Time (ignored, but required by `solve_ivp`).
-        w_flat: Flattened vorticity field of length ``n * n``.
-        n: Grid resolution per spatial axis.
-        dx: Grid spacing.
-        nu: Diffusion coefficient.
-        mu: Advection strength.
-        K3: Spectral multiplier used for the Poisson solve.
-
-    Returns:
-        Flattened ``dw/dt`` matching the shape of ``w_flat``.
-
-    Notes:
-        Implements ``dw/dt = nu * laplacian(w) - mu * (u * d/dx + v * d/dy)`` where
-        ``(u, v)`` is recovered from the streamfunction via ``u = dpsi/dy`` and
-        ``v = -dpsi/dx``.
-    """
-    w2d = w_flat.reshape(n, n)
-
-    # Streamfunction via spectral Poisson solver
-    psi = _spectral_poisson_solver(w2d, K3)
-
-    # Velocity field from psi
-    # u = dpsi/dy, v = -dpsi/dx
-    dpsidx, dpsidy = _gradient_periodic(psi, dx)  # dpsi/dx, dpsi/dy
-    u = dpsidy
-    v = -dpsidx
-
-    # Gradients of vorticity
-    dw_dx, dw_dy = _gradient_periodic(w2d, dx)
-
-    advec = u * dw_dx + v * dw_dy
-
-    diff = _laplacian_periodic(w2d, dx)
-
-    dwdt = nu * diff - mu * advec
-
-    return dwdt.ravel()
-
-
-def simulate_advection_diffusion(
-    x: NumpyLike,
-    return_timeseries: bool = False,
-    n: int = 50,
-    L: float = 10.0,
-    T: float = 80.0,
-    dt: float = 0.25,
-    integrator_kwargs: dict | None = None,
-) -> NumpyLike:
-    """Integrate the advection-diffusion system and emit physical channels.
-
-    Args:
-        x: Two-element vector ``[nu, mu]`` with viscosity and advection parameters.
-        return_timeseries: If ``True``, return the entire trajectory; otherwise only the
-            terminal state.
-        n: Grid resolution per spatial dimension.
-        L: Domain size along each axis.
-        T: End time for integration.
-        dt: Step between recorded solver outputs.
-        integrator_kwargs: Extra keyword arguments forwarded to `solve_ivp`.
-
-    Returns:
-        If `return_timeseries` is ``True``, an array of shape ``(n_time, n, n, 4)``;
-        otherwise shape ``(n, n, 4)``. Channels are ordered
-        ``[vorticity, u, v, streamfunction]``.
-    """
-    nu, mu = float(x[0]), float(x[1])
-
-    tspan = np.arange(0.0, T + 1e-12, dt)
-    n_time = len(tspan)
-
-    x_grid = np.linspace(-L / 2, L / 2, n, endpoint=False)
-    dx = float(x_grid[1] - x_grid[0])
-
-    # Initial condition: Gaussian vortex (centered)
-    X, Y = np.meshgrid(x_grid, x_grid)
-    w_initial = np.exp(-((X**2) + (Y**2) / 20.0))
-
-    # Precompute spectral operator for Poisson solve: 1/(k^2)
-    k = (2.0 * np.pi / L) * np.concatenate(
-        [np.arange(0, n // 2), np.arange(-n // 2, 0)]
-    )
-    KX, KY = np.meshgrid(k, k)
-    denom = KX**2 + KY**2
-    # Avoid division by zero at zero frequency: set to a large number then zero psi mean
-    denom[0, 0] = 1.0
-    K3 = 1.0 / denom
-    K3[0, 0] = 0.0
-
-    w0_flat = w_initial.ravel()
-
-    # Integrate
-    ode_kwargs = {**integrator_keywords, **(integrator_kwargs or {})}
-
-    sol = solve_ivp(
-        fun=lambda t, w: advection_diffusion_rhs(t, w, n, dx, nu, mu, K3),
-        t_span=(0.0, T),
-        y0=w0_flat,
-        t_eval=tspan,
-        **ode_kwargs,
-    )
-
-    if not sol.success:
-        raise RuntimeError("ODE solver failed: " + str(sol.message))
-
-    if return_timeseries:
-        # sol.y shape: (N, n_time)
-        w_ts = sol.y.T.reshape(n_time, n, n).astype(np.float32, copy=False)
-
-        channels = np.empty((n_time, n, n, 4), dtype=np.float32)
-        for ti, w2d in enumerate(w_ts):
-            psi = _spectral_poisson_solver(w2d, K3)
-            dpsidx, dpsidy = _gradient_periodic(psi, dx)
-            u = dpsidy
-            v = -dpsidx
-
-            channels[ti] = np.stack([w2d, u, v, psi], axis=-1).astype(
-                np.float32, copy=False
-            )
-
-        return channels
-
-    w_final = sol.y[:, -1].reshape(n, n)
-    psi = _spectral_poisson_solver(w_final, K3)
-    dpsidx, dpsidy = _gradient_periodic(psi, dx)
-    u = dpsidy
-    v = -dpsidx
-
-    return np.stack([w_final, u, v, psi], axis=-1).astype(np.float32, copy=False)
+from autosim.simulations.spatiotemporal.advection_diffusion_multichannel import *  # noqa: F403
