+++ src/autosim/simulations/spatiotemporal/__init__.py
@@ -1,11 +1,15 @@
 """Spatiotemporal simulator implementations."""
 
+from .advection_diffusion_multichannel import AdvectionDiffusionMultichannel
 from .gray_scott import GrayScott
 from .gross_pitaevskii import GrossPitaevskiiEquation2D
 from .navier_stokes_conditioned import ConditionedNavierStokes2D
+from .reaction_diffusion import ReactionDiffusion
 
 __all__ = [
+    "AdvectionDiffusionMultichannel",
     "ConditionedNavierStokes2D",
     "GrayScott",
     "GrossPitaevskiiEquation2D",
+    "ReactionDiffusion",
 ]
